fastapi-app/requirements.txt
# FastAPI e servidor web
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Prometheus metrics
prometheus-client==0.19.0

# OpenTelemetry - versões compatíveis
opentelemetry-distro[otlp]==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-instrumentation-requests==0.55b1

# Utilitários (opcional)
python-dotenv==1.0.0